+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Add `dynamicPositioning` prop to avoid the date popup appearing outside the screen (@stinger567)
+- `DateInput`: Avoid errors when `undefined` is passed as value
 
 ## 2.6.0 - 2023 Jul 11
 - Add 'yy' format (@erezarnon)